Enum lesspass_otp::Algorithm

pub enum Algorithm {
    SHA1,
    SHA256,
    SHA384,
    SHA512,
    SHA3_256,
    SHA3_384,
    SHA3_512,
}

Selects the hash algorithm to use in PBKDF or HMAC.

Variants

SHA1

SHA1.

Note: Using this algorithm does not work with LessPass.

SHA256

SHA2-256.

This is the algorithm used by the canonical LessPass implementation.

SHA384

SHA2-384.

Note: Using this algorithm makes the generated passwords different from every other LessPass implementation.

SHA512

SHA2-512.

Note: Using this algorithm makes the generated passwords different from every other LessPass implementation.

SHA3_256

SHA3-256.

Note: Using this algorithm makes the generated passwords different from every other LessPass implementation.

SHA3_384

SHA3-384.

Note: Using this algorithm makes the generated passwords different from every other LessPass implementation.

SHA3_512

SHA3-512.

Note: Using this algorithm makes the generated passwords different from every other LessPass implementation.

Implementations

impl Algorithm

#[must_use]pub fn pbkdf2(self, key: &[u8], data: &[u8], iterations: u32) -> Vec<u8>

Derive a PBKDF2 using current Algorithm.

The result length is variable:

20 chars for Algorithm::SHA1

32 chars for Algorithm::SHA256 or Algorithm::SHA3_256

48 chars for Algorithm::SHA384 or Algorithm::SHA3_384

64 chars for Algorithm::SHA512 or Algorithm::SHA3_512

Examples

use lesspass_otp::Algorithm;

let hash = Algorithm::SHA256.pbkdf2(b"myS3cre!K3y", b"Some salt", 1_000);
assert_eq!(hash, vec![
    227, 177, 151, 110, 153, 91, 123, 25, 111, 211, 151, 207, 114, 223,
    7, 194, 237, 243, 155, 62, 65, 201, 210, 230, 144, 213, 91, 151, 230,
    23, 64, 239
]);

let hash = Algorithm::SHA3_512.pbkdf2(b"myS3cre!K3y", b"Some salt", 1_000);
assert_eq!(hash, vec![
    233, 252, 44, 46, 18, 219, 245, 43, 176, 221, 248, 104, 5, 226, 170,
    242, 38, 161, 20, 240, 51, 167, 97, 138, 30, 222, 179, 48, 206, 169,
    56, 137, 247, 111, 153, 89, 58, 40, 209, 206, 153, 227, 100, 47, 222,
    255, 47, 158, 172, 175, 132, 171, 101, 109, 152, 167, 145, 232, 201,
    216, 2, 137, 139, 67
]);

#[must_use]pub fn hmac(self, key: &[u8], data: &[u8]) -> Vec<u8>

Derive a HMAC using current Algorithm.

The result length is variable:

20 chars for Algorithm::SHA1

32 chars for Algorithm::SHA256 or Algorithm::SHA3_256

48 chars for Algorithm::SHA384 or Algorithm::SHA3_384

64 chars for Algorithm::SHA512 or Algorithm::SHA3_512

Examples

use lesspass_otp::Algorithm;

let hash = Algorithm::SHA384.hmac(b"myS3cre!K3y", b"Some salt");
assert_eq!(hash, vec![
    101, 43, 178, 21, 155, 159, 249, 65, 0, 217, 135, 141, 114, 87, 92,
    89, 114, 74, 21, 79, 109, 214, 224, 231, 176, 95, 49, 94, 175, 109,
    87,82, 227, 88, 147, 14, 36, 84, 252, 11, 236, 112, 54, 245, 131,
    79, 184, 217
]);

let hash = Algorithm::SHA1.hmac(b"myS3cre!K3y", b"Some salt");
assert_eq!(hash, vec![
    232, 193, 68, 5, 132, 230, 202, 21, 208, 227, 112, 255, 88, 91,
    187, 37, 60, 193, 236, 34
]);