Function lesspass_otp::decode_base32

pub fn decode_base32(input: &str) -> Result<Vec<u8>, LessPassError>

Decode a base32 encoded string.

First, remove any number of = used for padding in input, then remove all - in the string, last remove all spaces before trying to decode the base32.

Examples

use lesspass_otp::decode_base32;

let base_32 = "JBSW-Y3DP-EBLW-64TM-MQQQ";
let decoded = decode_base32(base_32)?;
assert_eq!(&decoded, b"Hello World!");

Errors

Return LessPassError::InvalidBase32 if the input is not a valid base32 string.