Struct lesspass_otp::LessPass

pub struct LessPass<'a> { /* fields omitted */ }

The main struct, this is where we define the master password.

Implementations

impl<'a> LessPass<'a>

pub fn new(master: &'a str, algorithm: Algorithm) -> Result<Self, LessPassError>

Define master password to be used with every password.

The algorithm is the one used to generate the fingerprint, and the one used by default with password without algorithm specified.

Examples

use lesspass_otp::{Algorithm, LessPass};

let lp = LessPass::new("My5ecr3!", Algorithm::SHA256)?;

Errors

Could return a LessPassError::UnsupportedAlgorithm if the provided algorithm is not supported.

pub fn password(
&self,
site: &str,
login: &str,
counter: u32,
settings: &Settings
) -> Result<String, LessPassError>

Derive a password from the settings provided in the initialisation and identifications of the current site.

Examples

use lesspass_otp::{Algorithm, LessPass, Settings};

let lp = LessPass::new("My5ecr3!", Algorithm::SHA256)?;
// 16 characters, SHA2-256, lower, upper, number and symbol
let settings = Settings::default();

let pass = lp.password("example.com", "test@example.com", 1, &settings)?;
assert_eq!(pass, "38VdYgV3)/x*}`e,");

Errors

LessPassError::NoCharsetSelected if no characters are asked in the resulting password.

LessPassError::PasswordTooLong if the requested password length is too long for the current algorithm.

LessPassError::PasswordTooShort if the requested password is too short: less than 5 characters is forbidden.

LessPassError::UnsupportedAlgorithm in case you want to use an unsupported algorithm.

pub fn secret_hotp(
&self,
site: &str,
login: &str,
secret: &[u8]
) -> Result<Vec<u8>, LessPassError>

Decode a HOTP secret from aa previous encoded secret, or encode a clear one.

Note

This is not possible to encrypt a secret that is either 32 or 64 characters length, the secret will be considerated as encrypted and it will try to decrypt it.

Examples

use lesspass_otp::{Algorithm, decode_base32, LessPass, Settings};

let lp = LessPass::new("My5ecr3!", Algorithm::SHA256)?;
let settings = Settings::default();

// ----------------------
// Base32 decode the secret from the website
let secret = "JBSW Y3DP EBLW 64TM MQQQ";
let clear = decode_base32(secret).unwrap();
assert_eq!(clear, vec![72, 101, 108, 108, 111, 32, 87, 111, 114, 108, 100, 33]);

// Encrypt the secret
let encrypted_secret = lp.secret_hotp("example.com", "test@example.com", &clear)?;
assert_eq!(encrypted_secret, vec![
        101, 22, 162, 221, 2, 88, 94, 95, 176, 106, 204,
        94, 79, 92, 141, 190, 131, 49, 214, 61, 222, 201,
        120, 5, 188, 218, 35, 46, 210, 196, 21, 184
]);
// store the encrypted_secret anywhere, it cannot decrypted without master password
store_password(&encrypted_secret);

Decrypt the secret, then use it:

use lesspass_otp::{Algorithm, LessPass, Otp};

// Retrieve the encrypted password
let encrypted_secret = get_stored_encrypted_password();
// Initialise with the same master password
let lp = LessPass::new("My5ecr3!", Algorithm::SHA256)?;

// ----------------------
// Decrypt the stored encrypted secret
let clear_password = lp.secret_hotp("example.com", "test@example.com", &encrypted_secret)?;
assert_eq!(clear_password, vec![72, 101, 108, 108, 111, 32, 87, 111, 114, 108, 100, 33]);
// Use the clear_password with Otp::hotp in example
let otp = Otp::new(&clear_password, 6, None, None, None)?;
let token = otp.hotp(42);
assert_eq!(token, "063323");

Errors

Return the error LessPassError::InvalidLength if the secret is 0 or more than 64 characters length.

pub fn secret_totp(
&self,
site: &str,
login: &str,
secret: &[u8]
) -> Result<Vec<u8>, LessPassError>

Decode a TOTP secret from aa previous encoded secret, or encode a clear one.

Note

This is not possible to encrypt a secret that is either 32 or 64 characters length, the secret will be considerated as encrypted and it will try to decrypt it.

Examples

Encrypt the secret:

use lesspass_otp::{Algorithm, decode_base32, LessPass};

let lp = LessPass::new("My5ecr3!", Algorithm::SHA256)?;

// ----------------------
// Base32 decode the secret from the website
let secret = "JBSW Y3DP EBLW 64TM MQQQ";
let clear = decode_base32(secret).unwrap();
assert_eq!(clear, vec![72, 101, 108, 108, 111, 32, 87, 111, 114, 108, 100, 33]);

// Encrypt the secret
let encrypted_secret = lp.secret_totp("example.com", "test@example.com", &clear)?;
assert_eq!(encrypted_secret, vec![
        245, 248, 155, 215, 234, 198, 151, 5, 95, 75, 83,
        152, 159, 242, 191, 223, 59, 194, 6, 233, 107, 52,
        179, 27, 217, 250, 189, 86, 115, 118, 22, 138
]);
// store the encrypted_secret anywhere, it cannot be decrypted without master password
store_password(&encrypted_secret);

Decrypt the secret, then use it:

use lesspass_otp::{Algorithm, LessPass, Otp};

// Retrieve the encrypted password
let encrypted_secret = get_stored_encrypted_password();
// Initialise with the same master password
let lp = LessPass::new("My5ecr3!", Algorithm::SHA256)?;

// ----------------------
// Decrypt the stored encrypted secret
let clear_password = lp.secret_totp("example.com", "test@example.com", &encrypted_secret)?;
assert_eq!(clear_password, vec![72, 101, 108, 108, 111, 32, 87, 111, 114, 108, 100, 33]);
// Use the clear_password with Otp::totp in example
let otp = Otp::new(&clear_password, 6, None, None, None)?;
let token = otp.totp();

Errors

Return the error LessPassError::InvalidLength if the secret is 0 or more than 64 characters length.

#[must_use]pub fn get_fingerprint(&self, salt: &[u8]) -> [(&'static str, &'static str); 3]

Get master password fingerprint.

It contains an array of 3 symbols and 3 colors.

Examples

use lesspass_otp::{Algorithm, LessPass};

let lp = LessPass::new("My5ecr3!", Algorithm::SHA256)?;
let fingerprint = lp.get_fingerprint(b"");
assert_eq!(fingerprint, [
    ("#FF6CB6", "fa-beer"),
    ("#006CDB", "fa-hashtag"),
    ("#FFB5DA", "fa-cutlery"),
]);