Struct lesspass_otp::Settings

pub struct Settings { /* fields omitted */ }

Settings to derive a new password.

This is some common settings used to derive a new password.

Examples

use lesspass_otp::Settings;
use lesspass_otp::charset::{UpperCase, LowerCase, Symbols, Numbers};

// Create for a new password of 20 characters length, lower and uppercase characters and numbers
let settings = Settings::new(20, LowerCase::Using, UpperCase::Using, Numbers::Using, Symbols::NotUsing);

Implementations

impl Settings

#[must_use]pub fn new(
pass_len: u8,
lower: LowerCase,
upper: UpperCase,
num: Numbers,
sym: Symbols
) -> Self

Instantiate a new Settings, specifying the characters type and password length.

pub fn set_iterations(&mut self, iterations: u32)

Change number of iterations.

Notes

Doing so, your password will not be compatible anymore with stock Lesspass implementation.

Examples

use lesspass_otp::Settings;
use lesspass_otp::charset::{UpperCase, LowerCase, Symbols, Numbers};

// Create for a new password of 20 characters length, lower and uppercase characters and numbers
let mut settings = Settings::new(20, LowerCase::Using, UpperCase::Using, Numbers::Using, Symbols::NotUsing);
settings.set_iterations(20_000);

#[must_use]pub fn get_iterations(&self) -> u32

Get number of iterations configured, or default value.

#[must_use]pub const fn get_password_len(&self) -> u8

Get password length.

#[must_use]pub const fn get_characterset(&self) -> &CharacterSet

Retrieve configured CharacterSet.

pub fn set_algorithm(&mut self, algorithm: Algorithm)

Change default Algorithm.

Notes

Doing so, your password will not be compatible anymore with stock Lesspass implementation.

Examples

use lesspass_otp::{Settings, Algorithm};
use lesspass_otp::charset::{UpperCase, LowerCase, Symbols, Numbers};

// Create for a new password of 20 characters length, lower and uppercase characters and numbers
let mut settings = Settings::new(20, LowerCase::Using, UpperCase::Using, Numbers::Using, Symbols::NotUsing);
settings.set_algorithm(Algorithm::SHA512);

#[must_use]pub const fn get_algorithm(&self) -> Option<Algorithm>

Get the Algorithm.